+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `outerMargin` prop to `<BarChart/>` and `<MultiSeriesBarChart/>`
+
+### Changed
+
+- renamed `barOptions.margin` to `innerMargin` for `<BarChart/>` and `<MultiSeriesBarChart/>`
+
 ## [0.10.2] — 2021-05-05
 
 ### Fixed